Add files via upload
Roll the dice and rack up the points! Click hold to collect your points and give the other player a chance to roll. 
If you roll a 1 you lose your accumulated points and it goes to the next player.
So don't get too greedy! But don't be scared to get those points 😏. 1st too 100 wins and gets a special treat from me!